+++ include/mgos_mfrc522.h
@@ -10,7 +10,7 @@ void mgos_mfrc522_init(MFRC522 *rc5);
 void mgos_mfrc522_dump2serial(MFRC522 *rc5);
 bool mgos_mfrc522_isnewcardpresent(MFRC522 *rc5);
 bool mgos_mfrc522_readcardserial(MFRC522 *rc5);
-uint8_t mgos_mfrc522_readuic(MFRC522 *rc5,int i);
+int mgos_mfrc522_readuic(MFRC522 *rc5,int i);
 
 
 #ifdef __cplusplus
