+++ src/mgos_mfrc522.cpp
@@ -37,7 +37,7 @@ bool mgos_mfrc522_readcardserial(MFRC522 *rfid)
   return rfid->PICC_ReadCardSerial();
 }
 
-uint8_t mgos_mfrc522_readuic(MFRC522 *rfid, int i)
+int mgos_mfrc522_readuic(MFRC522 *rfid, int i)
  {
   if (rfid == nullptr) return 0;
   return rfid->PICC_getUID(i);
